feat : ProductList를 Class 형태로 생성
- productList를 Class 형태로 생성
